+++ pybamm/models/full_battery_models/lithium_ion/basic_dfn.py
@@ -282,3 +282,6 @@ class BasicDFN(BaseModel):
             pybamm.Event("Minimum voltage", voltage - param.voltage_low_cut),
             pybamm.Event("Maximum voltage", voltage - param.voltage_high_cut),
         ]
+
+    def new_copy(self, build=False):
+        return pybamm.BaseModel.new_copy(self)