+++ pybamm/solvers/casadi_solver.py
@@ -24,8 +24,6 @@ class CasadiSolver(pybamm.BaseSolver):
             - "safe": perform step-and-check integration in global steps of size \
             dt_max, checking whether events have been triggered. Recommended for \
             simulations of a full charge or discharge.
-            - "old safe": perform step-and-check integration in steps of size dt \
-            for each dt in t_eval, checking whether events have been triggered.
     rtol : float, optional
         The relative tolerance for the solver (default is 1e-6).
     atol : float, optional
@@ -73,15 +71,12 @@ class CasadiSolver(pybamm.BaseSolver):
         extra_options_call=None,
     ):
         super().__init__("problem dependent", rtol, atol, root_method, root_tol)
-        if mode in ["safe", "fast", "old safe"]:
+        if mode in ["safe", "fast"]:
             self.mode = mode
         else:
             raise ValueError(
-                """
-                invalid mode '{}'. Must be either 'safe' or 'old safe', for solving
-                with events, or 'fast', for solving quickly without events""".format(
-                    mode
-                )
+                "invalid mode '{}'. Must be 'safe', for solving with events, "
+                "or 'fast', for solving quickly without events".format(mode)
             )
         self.max_step_decrease_count = max_step_decrease_count
         self.dt_max = dt_max
@@ -92,9 +87,8 @@ class CasadiSolver(pybamm.BaseSolver):
         self.name = "CasADi solver with '{}' mode".format(mode)
 
         # Initialize
-        self.problems = {}
-        self.options = {}
-        self.methods = {}
+        self.integrators = {}
+        self.integrator_specs = {}
 
         pybamm.citations.register("Andersson2019")
 
@@ -111,22 +105,30 @@ class CasadiSolver(pybamm.BaseSolver):
         inputs : dict, optional
             Any external variables or input parameters to pass to the model when solving
         """
+        # Record whether there are any symbolic inputs
         inputs = inputs or {}
+        has_symbolic_inputs = any(isinstance(v, casadi.MX) for v in inputs.values())
+
         # convert inputs to casadi format
         inputs = casadi.vertcat(*[x for x in inputs.values()])
 
-        if self.mode == "fast":
-            integrator = self.get_integrator(model, t_eval, inputs)
-            solution = self._run_integrator(integrator, model, model.y0, inputs, t_eval)
+        if has_symbolic_inputs:
+            # Create integrax`tor without grid to avoid having to create several times
+            self.get_integrator(model, inputs)
+            solution = self._run_integrator(model, model.y0, inputs, t_eval)
             solution.termination = "final time"
             return solution
-        elif not model.events:
-            pybamm.logger.info("No events found, running fast mode")
-            integrator = self.get_integrator(model, t_eval, inputs)
-            solution = self._run_integrator(integrator, model, model.y0, inputs, t_eval)
+        elif self.mode == "fast" or not model.events:
+            if not model.events:
+                pybamm.logger.info("No events found, running fast mode")
+            # Create an integrator with the grid (we just need to do this once)
+            self.get_integrator(model, inputs, t_eval)
+            solution = self._run_integrator(model, model.y0, inputs, t_eval)
             solution.termination = "final time"
             return solution
         elif self.mode == "safe":
+            # Create integrator without grid to avoid having to create several times
+            self.get_integrator(model, inputs)
             y0 = model.y0
             if isinstance(y0, casadi.DM):
                 y0 = y0.full().flatten()
@@ -171,12 +173,12 @@ class CasadiSolver(pybamm.BaseSolver):
                     if len(t_window) == 1:
                         t_window = np.array([t, t + dt])
 
-                    integrator = self.get_integrator(model, t_window, inputs)
+                    # integrator = self.get_integrator(model, t_window, inputs)
                     # Try to solve with the current global step, if it fails then
                     # halve the step size and try again.
                     try:
                         current_step_sol = self._run_integrator(
-                            integrator, model, y0, inputs, t_window
+                            model, y0, inputs, t_window
                         )
                         solved = True
                     except pybamm.SolverError:
@@ -186,11 +188,11 @@ class CasadiSolver(pybamm.BaseSolver):
                     count += 1
                     if count >= self.max_step_decrease_count:
                         raise pybamm.SolverError(
-                            """
-                            Maximum number of decreased steps occurred at t={}. Try
-                            solving the model up to this time only or reducing dt_max.
-                            """.format(
-                                t
+                            "Maximum number of decreased steps occurred at t={}. Try "
+                            "solving the model up to this time only or reducing dt_max "
+                            "(currently, dt_max={})."
+                            "".format(
+                                t * model.timescale_eval, dt_max * model.timescale_eval
                             )
                         )
                 # Check most recent y to see if any events have been crossed
@@ -246,18 +248,25 @@ class CasadiSolver(pybamm.BaseSolver):
                     t_event = np.nanmin(t_events)
                     y_event = y_sol(t_event)
 
-                    # return truncated solution
-                    t_truncated = current_step_sol.t[current_step_sol.t < t_event]
-                    y_trunctaed = current_step_sol.y[:, 0 : len(t_truncated)]
-                    truncated_step_sol = pybamm.Solution(t_truncated, y_trunctaed)
-                    # assign temporary solve time
-                    truncated_step_sol.solve_time = np.nan
-                    # append solution from the current step to solution
-                    solution.append(truncated_step_sol)
+                    # solve again until the event time
+                    # See comments above on creating t_window
+                    t_window = np.concatenate(
+                        ([t], t_eval[(t_eval > t) & (t_eval < t_event)])
+                    )
+                    if len(t_window) == 1:
+                        t_window = np.array([t, t_event])
 
+                    # integrator = self.get_integrator(model, t_window, inputs)
+                    current_step_sol = self._run_integrator(model, y0, inputs, t_window)
+
+                    # assign temporary solve time
+                    current_step_sol.solve_time = np.nan
+                    # append solution from the current step to solution
+                    solution.append(current_step_sol)
                     solution.termination = "event"
                     solution.t_event = t_event
                     solution.y_event = y_event
+
                     break
                 else:
                     # assign temporary solve time
@@ -269,80 +278,29 @@ class CasadiSolver(pybamm.BaseSolver):
                     # update y0
                     y0 = solution.y[:, -1]
             return solution
-        elif self.mode == "old safe":
-            y0 = model.y0
-            if isinstance(y0, casadi.DM):
-                y0 = y0.full().flatten()
-            # Step-and-check
-            t = t_eval[0]
-            init_event_signs = np.sign(
-                np.concatenate(
-                    [event(t, y0, inputs) for event in model.terminate_events_eval]
-                )
-            )
-            pybamm.logger.info("Start solving {} with {}".format(model.name, self.name))
 
-            # Initialize solution
-            solution = pybamm.Solution(np.array([t]), y0[:, np.newaxis])
-            solution.solve_time = 0
-            for dt in np.diff(t_eval):
-                # Step
-                solved = False
-                count = 0
-                while not solved:
-                    integrator = self.get_integrator(
-                        model, np.array([t, t + dt]), inputs
-                    )
-                    # Try to solve with the current step, if it fails then halve the
-                    # step size and try again. This will make solution.t slightly
-                    # different to t_eval, but shouldn't matter too much as it should
-                    # only happen near events.
-                    try:
-                        current_step_sol = self._run_integrator(
-                            integrator, model, y0, inputs, np.array([t, t + dt])
-                        )
-                        solved = True
-                    except pybamm.SolverError:
-                        dt /= 2
-                    count += 1
-                    if count >= self.max_step_decrease_count:
-                        raise pybamm.SolverError(
-                            """
-                            Maximum number of decreased steps occurred at t={}. Try
-                            solving the model up to this time only.
-                            """.format(
-                                t
-                            )
-                        )
-                # Check most recent y
-                new_event_signs = np.sign(
-                    np.concatenate(
-                        [
-                            event(t, current_step_sol.y[:, -1], inputs)
-                            for event in model.terminate_events_eval
-                        ]
-                    )
-                )
-                # Exit loop if the sign of an event changes
-                if (new_event_signs != init_event_signs).any():
-                    solution.termination = "event"
-                    solution.t_event = solution.t[-1]
-                    solution.y_event = solution.y[:, -1]
-                    break
-                else:
-                    # assign temporary solve time
-                    current_step_sol.solve_time = np.nan
-                    # append solution from the current step to solution
-                    solution.append(current_step_sol)
-                    # update time
-                    t += dt
-                    # update y0
-                    y0 = solution.y[:, -1]
-            return solution
-
-    def get_integrator(self, model, t_eval, inputs):
+    def get_integrator(self, model, inputs, t_eval=None):
+        """
+        Method to create a casadi integrator object.
+        If t_eval is provided, the integrator uses t_eval to make the grid.
+        Otherwise, the integrator has grid [0,1].
+        """
+        # Use grid if t_eval is given
+        use_grid = not (t_eval is None)
         # Only set up problem once
-        if model not in self.problems:
+        if model in self.integrators:
+            # If we're not using the grid, we don't need to change the integrator
+            if use_grid is False:
+                return self.integrators[model]
+            # Otherwise, create new integrator with an updated (scaled) grid
+            else:
+                method, problem, options = self.integrator_specs[model]
+                t_eval_scaled = (t_eval - t_eval[0]) / (t_eval[-1] - t_eval[0])
+                options["grid"] = t_eval_scaled
+                integrator = casadi.integrator("F", method, problem, options)
+                self.integrators[model] = (integrator, use_grid)
+                return integrator
+        else:
             y0 = model.y0
             rhs = model.casadi_rhs
             algebraic = model.casadi_algebraic
@@ -358,51 +316,94 @@ class CasadiSolver(pybamm.BaseSolver):
 
             options = {
                 **self.extra_options_setup,
-                "grid": t_eval,
                 "reltol": self.rtol,
                 "abstol": self.atol,
-                "output_t0": True,
                 "show_eval_warnings": show_eval_warnings,
             }
 
             # set up and solve
             t = casadi.MX.sym("t")
             p = casadi.MX.sym("p", inputs.shape[0])
-            y_diff = casadi.MX.sym("y_diff", rhs(t_eval[0], y0, p).shape[0])
-            problem = {"t": t, "x": y_diff, "p": p}
-            if algebraic(t_eval[0], y0, p).is_empty():
+            y_diff = casadi.MX.sym("y_diff", rhs(0, y0, p).shape[0])
+
+            # rescale time
+            t_min = casadi.MX.sym("t_min")
+            t_max = casadi.MX.sym("t_max")
+            t_scaled = t_min + (t_max - t_min) * t
+            # add time limits as inputs
+            p_with_tlims = casadi.vertcat(p, t_min, t_max)
+
+            # save (scaled) grid
+            if use_grid is True:
+                t_eval_scaled = (t_eval - t_eval[0]) / (t_eval[-1] - t_eval[0])
+                options.update({"grid": t_eval_scaled, "output_t0": True})
+
+            problem = {"t": t, "x": y_diff, "p": p_with_tlims}
+            if algebraic(0, y0, p).is_empty():
                 method = "cvodes"
-                problem.update({"ode": rhs(t, y_diff, p)})
+                # rescale rhs by (t_max - t_min)
+                problem.update({"ode": (t_max - t_min) * rhs(t_scaled, y_diff, p)})
             else:
                 method = "idas"
-                y_alg = casadi.MX.sym("y_alg", algebraic(t_eval[0], y0, p).shape[0])
+                y_alg = casadi.MX.sym("y_alg", algebraic(0, y0, p).shape[0])
                 y_full = casadi.vertcat(y_diff, y_alg)
+                # rescale rhs by (t_max - t_min)
                 problem.update(
                     {
+                        "ode": (t_max - t_min) * rhs(t_scaled, y_full, p),
                         "z": y_alg,
-                        "ode": rhs(t, y_full, p),
-                        "alg": algebraic(t, y_full, p),
+                        "alg": algebraic(t_scaled, y_full, p),
                     }
                 )
-            self.problems[model] = problem
-            self.options[model] = options
-            self.methods[model] = method
-        else:
-            # problem stays the same
-            # just update options
-            self.options[model]["grid"] = t_eval
-        return casadi.integrator(
-            "F", self.methods[model], self.problems[model], self.options[model]
-        )
+            integrator = casadi.integrator("F", method, problem, options)
+            self.integrator_specs[model] = method, problem, options
+            self.integrators[model] = (integrator, use_grid)
+            return integrator
 
-    def _run_integrator(self, integrator, model, y0, inputs, t_eval):
-        rhs_size = model.concatenated_rhs.size
-        y0_diff, y0_alg = np.split(y0, [rhs_size])
+    def _run_integrator(self, model, y0, inputs, t_eval):
+        integrator, use_grid = self.integrators[model]
+        len_rhs = model.concatenated_rhs.size
+        y0_diff = y0[:len_rhs]
+        y0_alg = y0[len_rhs:]
         try:
             # Try solving
-            sol = integrator(x0=y0_diff, z0=y0_alg, p=inputs, **self.extra_options_call)
-            y_values = np.concatenate([sol["xf"].full(), sol["zf"].full()])
-            return pybamm.Solution(t_eval, y_values)
+            if use_grid is True:
+                t_min = t_eval[0]
+                t_max = t_eval[-1]
+                inputs_with_tlims = casadi.vertcat(inputs, t_min, t_max)
+                # Call the integrator once, with the grid
+                sol = integrator(
+                    x0=y0_diff,
+                    z0=y0_alg,
+                    p=inputs_with_tlims,
+                    **self.extra_options_call
+                )
+                y_sol = np.concatenate([sol["xf"].full(), sol["zf"].full()])
+                return pybamm.Solution(t_eval, y_sol)
+            else:
+                # Repeated calls to the integrator
+                x = y0_diff
+                z = y0_alg
+                y_diff = x
+                y_alg = z
+                for i in range(len(t_eval) - 1):
+                    t_min = t_eval[i]
+                    t_max = t_eval[i + 1]
+                    inputs_with_tlims = casadi.vertcat(inputs, t_min, t_max)
+                    sol = integrator(
+                        x0=x, z0=z, p=inputs_with_tlims, **self.extra_options_call
+                    )
+                    x = sol["xf"]
+                    z = sol["zf"]
+                    y_diff = casadi.horzcat(y_diff, x)
+                    if not z.is_empty():
+                        y_alg = casadi.horzcat(y_alg, z)
+                if z.is_empty():
+                    return pybamm.Solution(t_eval, y_diff)
+                else:
+                    y_sol = casadi.vertcat(y_diff, y_alg)
+                    return pybamm.Solution(t_eval, y_sol)
         except RuntimeError as e:
             # If it doesn't work raise error
             raise pybamm.SolverError(e.args[0])
+