+++ pybamm/solvers/processed_variable.py
@@ -254,31 +254,14 @@ class ProcessedVariable(object):
         """
         first_dim_nodes = self.mesh.nodes
         first_dim_edges = self.mesh.edges
-        second_dim_pts = self.base_variable.secondary_mesh.nodes
-        if self.base_eval.size // len(second_dim_pts) == len(first_dim_nodes):
+        second_dim_nodes = self.base_variable.secondary_mesh.nodes
+        second_dim_edges = self.base_variable.secondary_mesh.edges
+        if self.base_eval.size // len(second_dim_nodes) == len(first_dim_nodes):
             first_dim_pts = first_dim_nodes
-        elif self.base_eval.size // len(second_dim_pts) == len(first_dim_edges):
+        elif self.base_eval.size // len(second_dim_nodes) == len(first_dim_edges):
             first_dim_pts = first_dim_edges
 
-        # Set spatial variable names
-        first_spatial_variable_names = self.mesh.spatial_variable_names
-        if len(first_spatial_variable_names) != 1:
-            raise ValueError("Found too many spatial variables for a 1D variable")
-        else:
-            first_spatial_variable = first_spatial_variable_names[0]
-            self.first_dimension = first_spatial_variable
-            setattr(self, first_spatial_variable + "_sol", first_dim_pts)
-
-        second_spatial_variable_names = (
-            self.base_variable.secondary_mesh.spatial_variable_names
-        )
-        if len(second_spatial_variable_names) != 1:
-            raise ValueError("Found too many spatial variables for a 1D variable")
-        else:
-            second_spatial_variable = second_spatial_variable_names[0]
-            self.second_dimension = second_spatial_variable
-            setattr(self, second_spatial_variable + "_sol", second_dim_pts)
-
+        second_dim_pts = second_dim_nodes
         first_dim_size = len(first_dim_pts)
         second_dim_size = len(second_dim_pts)
         entries = np.empty((first_dim_size, second_dim_size, len(self.t_sol)))
@@ -305,27 +288,97 @@ class ProcessedVariable(object):
                     order="F",
                 )
 
+        # add points outside first dimension domain for extrapolation to
+        # boundaries
+        extrap_space_first_dim_left = np.array(
+            [2 * first_dim_pts[0] - first_dim_pts[1]]
+        )
+        extrap_space_first_dim_right = np.array(
+            [2 * first_dim_pts[-1] - first_dim_pts[-2]]
+        )
+        first_dim_pts = np.concatenate(
+            [extrap_space_first_dim_left, first_dim_pts, extrap_space_first_dim_right]
+        )
+        extrap_entries_left = np.expand_dims(2 * entries[0] - entries[1], axis=0)
+        extrap_entries_right = np.expand_dims(2 * entries[-1] - entries[-2], axis=0)
+        entries_for_interp = np.concatenate(
+            [extrap_entries_left, entries, extrap_entries_right], axis=0
+        )
+
+        # add points outside second dimension domain for extrapolation to
+        # boundaries
+        extrap_space_second_dim_left = np.array(
+            [2 * second_dim_pts[0] - second_dim_pts[1]]
+        )
+        extrap_space_second_dim_right = np.array(
+            [2 * second_dim_pts[-1] - second_dim_pts[-2]]
+        )
+        second_dim_pts = np.concatenate(
+            [
+                extrap_space_second_dim_left,
+                second_dim_pts,
+                extrap_space_second_dim_right,
+            ]
+        )
+        extrap_entries_second_dim_left = np.expand_dims(
+            2 * entries_for_interp[:, 0, :] - entries_for_interp[:, 1, :], axis=1
+        )
+        extrap_entries_second_dim_right = np.expand_dims(
+            2 * entries_for_interp[:, -1, :] - entries_for_interp[:, -2, :], axis=1
+        )
+        entries_for_interp = np.concatenate(
+            [
+                extrap_entries_second_dim_left,
+                entries_for_interp,
+                extrap_entries_second_dim_right,
+            ],
+            axis=1,
+        )
+
+        # Set spatial variable names
+        first_spatial_variable_names = self.mesh.spatial_variable_names
+        if len(first_spatial_variable_names) != 1:
+            raise ValueError("Found too many spatial variables for a 1D variable")
+        else:
+            first_spatial_variable = first_spatial_variable_names[0]
+            self.first_dimension = first_spatial_variable
+            setattr(self, first_spatial_variable + "_sol", first_dim_pts)
+
+        second_spatial_variable_names = (
+            self.base_variable.secondary_mesh.spatial_variable_names
+        )
+        if len(second_spatial_variable_names) != 1:
+            raise ValueError("Found too many spatial variables for a 1D variable")
+        else:
+            second_spatial_variable = second_spatial_variable_names[0]
+            self.second_dimension = second_spatial_variable
+            setattr(self, second_spatial_variable + "_sol", second_dim_pts)
+
         # assign attributes for reference
         self.entries = entries
         self.dimensions = 2
         first_length_scale = self.get_spatial_scale(
             self.first_dimension, self.domain[0]
         )
-        self.first_dim_pts = first_dim_pts * first_length_scale
+        first_dim_pts_for_interp = first_dim_pts * first_length_scale
 
         second_length_scale = self.get_spatial_scale(
             self.second_dimension, self.auxiliary_domains["secondary"][0]
         )
-        self.second_dim_pts = second_dim_pts * second_length_scale
+        second_dim_pts_for_interp = second_dim_pts * second_length_scale
+
+        # Set pts to edges for nicer plotting
+        self.first_dim_pts = first_dim_edges * first_length_scale
+        self.second_dim_pts = second_dim_edges * second_length_scale
 
         # set up interpolation
         if len(self.t_sol) == 1:
             # function of space only. Note the order of the points is the reverse
             # of what you'd expect
             interpolant = interp.interp2d(
-                self.second_dim_pts,
-                self.first_dim_pts,
-                entries[:, :, 0],
+                second_dim_pts_for_interp,
+                first_dim_pts_for_interp,
+                entries_for_interp[:, :, 0],
                 kind="linear",
                 fill_value=np.nan,
                 bounds_error=False,
@@ -338,8 +391,8 @@ class ProcessedVariable(object):
         else:
             # function of space and time.
             self._interpolation_function = interp.RegularGridInterpolator(
-                (self.first_dim_pts, self.second_dim_pts, self.t_pts),
-                entries,
+                (first_dim_pts_for_interp, second_dim_pts_for_interp, self.t_pts),
+                entries_for_interp,
                 method="linear",
                 fill_value=np.nan,
                 bounds_error=False,